🚚 usePulldown() 관리파일 이동
* issue: [NAILCASE-54](https://nailcase.atlassian.net/browse/NAILCASE-54)
desc:
- 컨밴션에맞게 use-common.ts => use-component.ts로 관리파일이동
- 공용컴포넌트 hook은 use-component.ts에서 관리
